# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gym Ledger API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Login user",
        "description": "Authenticate user and return JWT token stored in cookie",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "ateeb@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": 1234567890
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "example": "Login successful"
                    },
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "example": "Ateeb"
                        },
                        "email": {
                          "type": "string",
                          "example": "ateeb@gmail.com"
                        },
                        "role": {
                          "type": "integer",
                          "example": 0
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials or missing fields"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/auth/signup": {
      "post": {
        "summary": "Register new user",
        "description": "Create a new user account",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password",
                  "confirmPass"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Ateeb"
                  },
                  "email": {
                    "type": "string",
                    "example": "ateeb@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": 12345678
                  },
                  "confirmPass": {
                    "type": "string",
                    "example": 12345678
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "example": "User registered successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or user already exists"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/members/add_member": {
      "post": {
        "summary": "Add new gym member",
        "description": "Create a new gym member under the logged-in admin",
        "tags": [
          "Members"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "contact",
                  "emergency_contact",
                  "height",
                  "weight",
                  "fitness_goal",
                  "personal_training",
                  "assign_trainer",
                  "medical_condition"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "contact": {
                    "type": "string"
                  },
                  "emergency_contact": {
                    "type": "string"
                  },
                  "height": {
                    "type": "number"
                  },
                  "weight": {
                    "type": "number"
                  },
                  "fitness_goal": {
                    "type": "string"
                  },
                  "personal_training": {
                    "type": "boolean"
                  },
                  "assign_trainer": {
                    "type": "string"
                  },
                  "medical_condition": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Member created successfully"
          },
          "400": {
            "description": "Validation error or member already exists"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}